To compare sub-regression structures

Compares two sub-regression structures, considering one of them as the "true one".

compare_struct(trueZ = trueZ, Zalgo = Zalgo, all = TRUE, mode = "NULL")

Arguments

trueZ
first structure (binary adjacency matrix)

Zalgo
second structure (binary adjacency matrix)

all
(boolean) Also compute the ratio for each stat.

mode
how to modify the structures before comparison. mode=c("NULL","hybrid","clique","sym") It allows to compare groups instead of exact sub-regressions. Does nothing by default.

Value

true1

Number of links that exist in both matrices

false1

Number of links that exist only in Zalgo

false0

Number of links that exist only in trueZ

deltadr

Number of sub-regressions in trueZ - Number of sub-regressions in Zalgo (i.e.: negative if too much sub-regressions in Zalgo)

true_left

Number of variables redundant in both matrices

false_left

Number of variables redundant in Zalgo but not in trueZ

ratio_true1

ratio of links in trueZ that exist also in Zalgo

ratio_true0

ratio of links not in trueZ that don't exist in Zalgo.